refactor: hoist isPrivateNetworkEndpoint into server/lib/localEndpoint.js (#6818)
isPrivateNetworkEndpoint existed twice — server/lib/providerPrerequisites.js
and client/src/utils/providerEndpoints.js — and the two answered two
documented edge cases differently (a compact http://127.1 literal, a
bracketless IPv6 literal), with only a comment noting the divergence. Move
the function and its helpers (endpointHost, isIpv4Literal, isPrivateIpv6,
PRIVATE_IP_RE) from providerPrerequisites.js into the existing
server/lib/localEndpoint.js leaf, which already held the related
isLocalInstanceHost/isLocalInstanceEndpoint predicates; providerPrerequisites.js
now imports it (restoring the local binding isFleetProvider-adjacent code
needs) instead of defining its own. The client re-exports the same leaf, so
the server's answer — the stricter, URL-parsing one — now wins for both.

isLocalEndpoint, isLocalInstanceProvider, and localBackendForProvider stay
client-side: they are deliberately separate rules (a cheaper regex, a
blank-endpoint-is-local default, a cheaper backend labeler), not hand copies
of a server algorithm.
